bg doctor: name the failing agent, and cast the interval binds
Two things, one of which nearly shipped broken.

doctor reported "44 failures" — a number an operator can do nothing with.
It now names the agent and the reason: "skein 12 of 14 calls — ran out of
room before it finished writing its answer", using the same classifier as
/flock and the API. It also says when no daily token ceiling is set, which
is the state production has been in all along: the only thing between the
newsroom and a 48-minute lockout was luck.

And the cast. make_interval has no bigint overload, so an i64 bind arrives
as a type it will not accept and the whole query errors. Both queries added
this session had it — failure_rates and the declines upsert — and both
callers swallow errors, so the check that exists to make failures visible
would itself have failed invisibly, and the backoff would have recorded
nothing while appearing to work. Found by running the command rather than
by reading the diff. The gaggles queries already cast to i32, so the trap
was known; it is now commented where it bites.

+++ crates/bg-db/src/agents.rs
@@ -330,7 +330,11 @@ pub async fn failure_rates(
                (array_agg(error order by started_at desc)
                     filter (where status = 'failed' and error is not null))[1]
         from agent_runs
-        where started_at > now() - make_interval(hours => $1)
+        -- `::int` is load-bearing: make_interval has no bigint overload, and an
+        -- i64 bind arrives as bigint. Without the cast this query errors, the
+        -- caller's `if let Ok` swallows it, and the check reports nothing at
+        -- all — which is indistinguishable from "everything is fine".
+        where started_at > now() - make_interval(hours => $1::int)
         group by role
         order by count(*) desc
         "#,
